refactor(suite-common): revert yield exports
