+++ packages/eslint/src/localRulesConfig.mjs
@@ -33,6 +33,8 @@ const publishableTrezorPackages = [
     '@suite-common/schemas',
 ];
 
+const packagesWithSectionEntryPoints = ['@suite-common/earn-stablecoin'];
+
 /** @type {Config[]} */
 export const localRulesConfig = [
     {
@@ -57,7 +59,10 @@ export const localRulesConfig = [
                 'error',
                 {
                     packageScopes: ['@suite-native', '@suite', '@suite-common', '@trezor'],
-                    ignoredPackages: publishableTrezorPackages,
+                    ignoredPackages: [
+                        ...publishableTrezorPackages,
+                        ...packagesWithSectionEntryPoints,
+                    ],
                 },
             ],
             'local-rules/analytics-event-name': 'error',
