+++ packages/suite/src/components/tx-simulation/earn-stablecoin/EarnYieldTxSimulationModal.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 
 import { type UserContextModalType } from '@suite/modal';
-import { composeStablecoinYieldTxSimulationAction } from '@suite-common/earn-stablecoin';
+import { composeStablecoinYieldTxSimulationAction } from '@suite-common/earn-stablecoin/src/tx-simulation';
 import { selectAccountByKey } from '@suite-common/wallet-core';
 
 import { useSelector } from 'src/hooks/suite';
