+++ suite-common/earn-stablecoin/package.json
@@ -5,7 +5,6 @@
     "license": "See LICENSE.md in repo root",
     "sideEffects": false,
     "type": "module",
-    "main": "src/index",
     "scripts": {
         "depcheck": "yarn g:depcheck",
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
