--- suite-common/earn-stablecoin/src/index.ts
@@ -1,2 +1,0 @@
-export * from './tx-simulation';
-export * from './signing';
