+++ suite-native/module-earn/src/yieldTransactionThunks.ts
@@ -1,5 +1,5 @@
 import { selectSelectedDevice } from '@suite-common/device';
-import { buildStablecoinYieldTransactionReview } from '@suite-common/earn-stablecoin';
+import { buildStablecoinYieldTransactionReview } from '@suite-common/earn-stablecoin/src/signing';
 import { createThunk } from '@suite-common/redux-utils';
 import {
     type YieldFlowResolvedData,
